+++ src/main/java/kr/co/iteyes/agenthelper/entity/id/AllergyId.java
@@ -5,11 +5,12 @@ import lombok.NoArgsConstructor;
 
 import javax.persistence.Id;
 import java.io.Serializable;
+import java.math.BigInteger;
 import java.sql.Timestamp;
 
 @EqualsAndHashCode
 @NoArgsConstructor
 public class AllergyId implements Serializable {
-    private Long seqNo;
+    private BigInteger seqNo;
     private String patId;
 }
