+++ src/main/java/kr/co/iteyes/agenthelper/entity/id/EncounterId.java
@@ -5,11 +5,12 @@ import lombok.NoArgsConstructor;
 
 import javax.persistence.Id;
 import java.io.Serializable;
+import java.math.BigInteger;
 import java.sql.Timestamp;
 
 @EqualsAndHashCode
 @NoArgsConstructor
 public class EncounterId implements Serializable {
-    private Long rcptNo;
+    private BigInteger rcptNo;
     private Timestamp lastChgDt;
 }
