+++ src/main/java/kr/co/iteyes/agenthelper/entity/id/MediaId.java
@@ -5,12 +5,13 @@ import lombok.NoArgsConstructor;
 
 import javax.persistence.Id;
 import java.io.Serializable;
+import java.math.BigInteger;
 import java.sql.Date;
 import java.sql.Timestamp;
 
 @EqualsAndHashCode
 @NoArgsConstructor
 public class MediaId implements Serializable {
-    private Long seqNo;
+    private BigInteger seqNo;
     private String patId;
 }
