+++ src/main/java/kr/co/iteyes/agenthelper/entity/id/MedicationId.java
@@ -6,13 +6,14 @@ import lombok.NoArgsConstructor;
 import javax.persistence.Column;
 import javax.persistence.Id;
 import java.io.Serializable;
+import java.math.BigInteger;
 import java.sql.Timestamp;
 
 @EqualsAndHashCode
 @NoArgsConstructor
 public class MedicationId implements Serializable {
-    private Long seqNo;
-    private Long rcptNo;
+    private BigInteger seqNo;
+    private BigInteger rcptNo;
     private String patId;
     private Timestamp lastChgDt;
 }
