+++ src/main/java/kr/co/iteyes/agenthelper/repository/DiagnosticRadioRepository.java
@@ -5,10 +5,11 @@ import kr.co.iteyes.agenthelper.entity.id.DiagnosticRadioId;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
+import java.math.BigInteger;
 import java.util.Date;
 import java.util.List;
 
 public interface DiagnosticRadioRepository extends JpaRepository<DiagnosticRadio, DiagnosticRadioId> {
     @Query("select d from DiagnosticRadio d where d.patId = :patientId and d.rcptNo = :rcptNo")
-    List<DiagnosticRadio> findAllResource(String patientId, Long rcptNo);
+    List<DiagnosticRadio> findAllResource(String patientId, BigInteger rcptNo);
 }
