+++ src/main/java/kr/co/iteyes/agenthelper/repository/ObservationFuncRepository.java
@@ -5,10 +5,11 @@ import kr.co.iteyes.agenthelper.entity.id.ObservationFuncId;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
+import java.math.BigInteger;
 import java.util.Date;
 import java.util.List;
 
 public interface ObservationFuncRepository extends JpaRepository<ObservationFunc, ObservationFuncId> {
     @Query("select o from ObservationFunc o where o.patId = :patientId and o.rcptNo = :rcptNo")
-    List<ObservationFunc> findAllResource(String patientId, Long rcptNo);
+    List<ObservationFunc> findAllResource(String patientId, BigInteger rcptNo);
 }
